requirements.txt
pytimeparse==1.1.*
alpha-vantage==2.1.*
matplotlib==3.1.*
numpy==1.18.*
pandas==0.25.*
python-telegram-bot==12.2.*